refactor: move error 404 component
