+++ app/src/App.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom'
 import { WorkspaceSelection } from './features/workspace'
 import { useStore } from './data/store'
-import Error404 from './components/Error404'
+import { Error404 } from './features/ui'
 import ImprovementsPage from './pages/ImprovementsPage'
 import IssuesPage from './pages/IssuesPage'
 import IssuePage from './pages/IssuePage'
