+++ app/src/features/ui/index.ts
@@ -2,3 +2,4 @@ export * from './Narrow'
 export * from './AttributionCard'
 export * from './InfoBubble'
 export * from './AvatarDiv'
+export * from './Error404'
