+++ app/src/pages/ImprovementPage.tsx
@@ -5,7 +5,7 @@ import { IMPROVEMENT_STATUS_UI } from '../components/Status'
 import { renderDate } from '../data/util'
 import { ImprovementIcon } from '../components/ItemIcon'
 import { PageHeader } from '../components/PageHeader'
-import Error404 from '../components/Error404'
+import { Error404 } from '../features/ui'
 import { useParams } from 'react-router-dom'
 import { TicketLink } from '../components/TicketLink'
 import { ImprovementId } from '../data/types'
